fix: move methods imported by UI from test utils back to src utils
